+++ scripts/release.sh
@@ -21,3 +21,17 @@ cargo build --release --target x86_64-unknown-linux-musl
 
 # Pack the binary to reduce size
 upx --best --lzma ./target/x86_64-unknown-linux-musl/release/${pkg_name}
+
+# Copy binary
+cp ./target/x86_64-unknown-linux-musl/release/${pkg_name} "${release_path}"/usr/local/bin
+# Generate DEBIAN controll
+cat > "${release_path}"/DEBIAN/control <<- EOM
+Package: ${pkg_name}
+Version: ${version}
+Architecture: ${arch}
+Maintainer: Next Hat team@next-hat.com
+Description: Convert docker-compose.yaml to Nanocl StateFile.yml
+EOM
+
+mkdir -p ./target/debian
+dpkg-deb --build --root-owner-group "${release_path}" ./target/debian/${pkg_name}_"${version}"_"${arch}".deb