+++ changelog.md
@@ -9,4 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Convert `docker-compose.yml` to `StateFile.yml`
+- Convert `docker-compose.yml` to Nanocl `StateFile.yml`